+++ NEWS.md
@@ -1,6 +1,7 @@
 # gedcomS7 0.2.0 (XX Sep 2025)
 
 * `DateGregorian()` now handles BCE dates with a negative `@year`
+* GEDCOM files within zip archives can be imported (GEDzip)
 * Minor changes to some internal class names and their inheritance
 * Other minor internal improvements